refactor: refactor and move the scripts around
