--- docs/API_reference/extract_variables_from_files.md
@@ -1,10 +1,0 @@
-<!--
-SPDX-FileCopyrightText: 2025 GFZ Helmholtz Centre for Geosciences
-SPDX-FileContributor: Bernhard Haas
-
-SPDX-License-Identifier: Apache-2.0
--->
-
-::: el_paso.extract_variables_from_files.extract_variables_from_files
-
-::: el_paso.extract_variables_from_files.ExtractionInfo
