--- docs/API_reference/processing/convert_string_to_datetime.md
@@ -1,8 +1,0 @@
-<!--
-SPDX-FileCopyrightText: 2025 GFZ Helmholtz Centre for Geosciences
-SPDX-FileContributor: Bernhard Haas
-
-SPDX-License-Identifier: Apache-2.0
--->
-
-::: el_paso.processing.convert_string_to_datetime.convert_string_to_datetime
