+++ docs/API_reference/processing/extraction.md
@@ -1,0 +1,10 @@
+<!--
+SPDX-FileCopyrightText: 2025 GFZ Helmholtz Centre for Geosciences
+SPDX-FileContributor: Bernhard Haas
+
+SPDX-License-Identifier: Apache-2.0
+-->
+
+::: el_paso.processing.extraction.extract_variables_from_files
+
+::: el_paso.processing.extraction.ExtractionInfo
