+++ docs/API_reference/utilities/convert_string_to_datetime.md
@@ -1,0 +1,8 @@
+<!--
+SPDX-FileCopyrightText: 2025 GFZ Helmholtz Centre for Geosciences
+SPDX-FileContributor: Bernhard Haas
+
+SPDX-License-Identifier: Apache-2.0
+-->
+
+::: el_paso.utils.convert_string_to_datetime
