+++ el_paso/__init__.py
@@ -25,8 +25,8 @@ from el_paso import physics, processing, saving_strategies, units, data_standard
 from el_paso.save import save
 from el_paso.processing import TimeBinMethod
 from el_paso.download import download
-from el_paso.extract_variables_from_files import extract_variables_from_files, ExtractionInfo
 from el_paso.load_indices_solar_wind_parameters import load_indices_solar_wind_parameters
+from el_paso import utils
 
 # expose RBMDataSet related classes and functions
 from swvo.io.RBMDataSet.custom_enums import (
@@ -58,13 +58,11 @@ __all__ = [
     "IRBEM_SYSAXIS_GSM",
     "IRBEM_SYSAXIS_MAG",
     "IRBEM_SYSAXIS_SM",
-    "ExtractionInfo",
     "TimeBinMethod",
     "Variable",
     "activate_release_mode",
     "data_standards",
     "download",
-    "extract_variables_from_files",
     "get_release_msg",
     "is_in_release_mode",
     "load_indices_solar_wind_parameters",
@@ -73,6 +71,7 @@ __all__ = [
     "save",
     "saving_strategies",
     "units",
+    "utils",
 ]
 
 __version__ = "1.0.3rc0"
