+++ examples/Arase/arase_mepe.py
@@ -65,31 +65,31 @@ def process_mepe_level_3(  # noqa: PLR0915
     )
 
     extraction_infos = [
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Epoch",
             name_or_column="epoch",
             unit=ep.units.tt2000,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Energy",
             name_or_column="FEDU_Energy",
             unit=u.keV,
             is_time_dependent=False,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Pitch_angle",
             name_or_column="FEDU_Alpha",
             unit=u.deg,
             is_time_dependent=False,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="FEDU",
             name_or_column="FEDU",
             unit=(u.cm**2 * u.s * u.sr * u.keV) ** (-1),
         ),
     ]
 
-    mepe_variables = ep.extract_variables_from_files(
+    mepe_variables = ep.processing.extract_variables_from_files(
         start_time,
         end_time,
         "daily",
