+++ examples/Arase/get_arase_orbit_variables.py
@@ -37,19 +37,19 @@ def get_arase_orbit_level_2_variables(
     )
 
     extraction_infos = [
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Epoch",
             name_or_column="epoch",
             unit=ep.units.tt2000,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="pos_sm",
             name_or_column="pos_sm",
             unit=u.dimensionless_unscaled,
         ),
     ]
 
-    variables = ep.extract_variables_from_files(
+    variables = ep.processing.extract_variables_from_files(
         start_time,
         end_time,
         "daily",
@@ -108,39 +108,39 @@ def get_arase_orbit_level_3_variables(
             raise ValueError(msg)
 
     extraction_infos = [
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Epoch",
             name_or_column="epoch",
             unit=ep.units.tt2000,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="B_local",
             name_or_column=f"pos_blocal_{mag_field_label}",
             unit=u.nT,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="B_eq",
             name_or_column=f"pos_beq_{mag_field_label}",
             unit=u.nT,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Lm",
             name_or_column=f"pos_lmc_{mag_field_label}",
             unit=u.dimensionless_unscaled,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="Lstar",
             name_or_column=f"pos_lstar_{mag_field_label}",
             unit=u.dimensionless_unscaled,
         ),
-        ep.ExtractionInfo(
+        ep.processing.ExtractionInfo(
             result_key="pos_eq",
             name_or_column=f"pos_eq_{mag_field_label}",
             unit=u.dimensionless_unscaled,
         ),
     ]
 
-    variables = ep.extract_variables_from_files(
+    variables = ep.processing.extract_variables_from_files(
         start_time,
         end_time,
         "daily",
